+++ CHANGELOG.md
@@ -9,10 +9,13 @@ glance - what to expact from upgrading to a new version.
 ### Added
 
 - Per-user and global rate limits.
+- `!hive` as alias to `!hive2`
+- Description of bot commands to help output.
 
 ### Changed
 
 - Improved formatting of Hive 2 player data.
+- `!hive` command defaults to Discord username if not specified otherwise.
 
 ### Fixed